fix: 修正 DiscountTest 的當前時間與 discount 型態
調整項目:
1. 修復 DiscountTest 中, 當前時間可能不符合營業時間
2. 修復 discount 型態為 double

+++ src/main/java/main/Discount.java
@@ -1,7 +1,7 @@
 package main;
 
 public class Discount {
-	private float discount = 0;
+	private double discount = 0;
 	
 	public Discount(Identity identity, String dateTime) {
 
@@ -14,20 +14,20 @@ public class Discount {
 			System.out.println("Your age doesn't meet the requirements.");
 		} else {
 			if (identity.isMember()) {
-				discount = (float) 0.5;
+				discount = 0.5;
 			} else if (identity.isGroup()) {
-				discount = (float) 0.7;
+				discount = 0.7;
 			} else if (identity.getAge() < 12 || identity.getAge() >= 60) {
-				discount = (float) 0.8;
+				discount = 0.8;
 			} else if (hour >= 5 && hour < 7) {
-				discount = (float) 0.8;
+				discount = 0.8;
 			} else {
 				discount = 1;
 			}
 		}
 	}
 
-	public float getDiscount() {
+	public double getDiscount() {
 		return discount;
 	}
 }
